backend/requirements.txt
dash
dash-bootstrap-components
fastapi
uvicorn
psycopg2-binary
python-dotenv
apscheduler
openai
anthropic
google-genai
pydantic[email]
pydantic-settings
requests
passlib[bcrypt]
bcrypt==3.2.0
python-jose[cryptography]
python-multipart
plotly
pandas

